Only use RDiscount, don't try to use other Markdown libraries
The three Markdown libraries supported are not very compatible, which leads
to problems with inconsistent formatting or simply failing to work at
all. Bluecloth does not support the :smart option, while Redcarpet just
ignores it.  Redcarpet does not parse the Markdown generated by dockblocks
propererly.  Since these three libraries are not compatible, it doesn't
seem to make sense to try to support all of them, especially since we
can just specify in our gem the dependency on the one that works best.

Fixes: https://github.com/rtomayko/rocco/issues/74 (fix rtomayko/rocco#74)